pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["mccole"]

[tool.setuptools.package-data]
mypkg = ["*.css", "*.js", "*.html"]

[project]
name = "mccole"
version = "0.4.1"
authors = [
    {name = "Greg Wilson", email = "gvwilson@third-bit.com"},
    {name = "Juanan Pereira", email = "juanan.pereira@ehu.eus"},
]
maintainers = [
    {name = "Greg Wilson", email = "gvwilson@third-bit.com"},
]
description = "Simple template for static web sites"
keywords = ["static site generator"]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "beautifulsoup4>=4.12.3",
    "esprima>=4.0.1",
    "gitpython>=3.1",
    "html5validator>=0.4.2",
    "jinja2>=3.1.4",
    "markdown>=3.6",
    "prettytable>=3.11",
    "pygithub>=2.4",
    "pyyaml>=6.0",
    "ruff>=0.6.0",
    "tomli>=2.0.1",
]

[project.optional-dependencies]
docs = [
  "markdown-include",
  "mkdocs",
  "mkdocstrings[python]",
]

[project.scripts]
mccole = "mccole:clui.main"

[tool.mccole]
skips = ["dist", "mccole.egg-info"]

[tool.mccole.renames]
"CODE_OF_CONDUCT.md" = "code_of_conduct.md"
"CONTRIBUTING.md" = "contributing.md"
"LICENSE.md" = "license.md"
"README.md" = "index.md"
